Clean application of proposed changes as well as private renaming
The private renaming is mostly because I keep confusing the renamed element, and this helps me.
It fits the rename from wgsl_bindgen. To be accepted upstream it would likely have to be backwards compatible,
which might be achievable with type alias + deprecated notice, but we will consider that when we get there.

Other than that, this allows reuse of generated code for a more complex application, by allowing access to innards.